requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
statsmodels
Flask
ipykernel
django
djangorestframework